Update opentelemetry-collector to v0.76.1 (#31)
* Update opentelemetry-collector

Some tests needed to be adapted due to
grpc codes no longer being converted to
strings. Turns out they should be integers
anyway, but I've left the existing string
handling of rpc.grpc.status_code for now.

* Bump Go to 1.19

* Update staticcheck

+++ input/otlp/traces_test.go
@@ -1377,13 +1377,13 @@ func TestGRPCTransactionFromNodejsSDK(t *testing.T) {
 			assert.Equal(t, "request", event.Transaction.Type)
 		}
 		test(t, map[string]interface{}{
-			"rpc.grpc.status_code": codes.Unavailable,
+			"rpc.grpc.status_code": int64(codes.Unavailable),
 		})
 	})
 
 	t.Run("span transformation", func(t *testing.T) {
 		event := transformSpanWithAttributes(t, map[string]interface{}{
-			"rpc.grpc.status_code": codes.Unavailable,
+			"rpc.grpc.status_code": int64(codes.Unavailable),
 		})
 		assert.Equal(t, "external", event.Span.Type)
 		assert.Equal(t, "grpc", event.Span.Subtype)
